Django API/myproject/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
